perf: :zap: only smolsort_3 when only 3 elems on  stack in recursion
