+++ makefile
@@ -6,7 +6,7 @@
 #    By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/09/09 14:12:20 by youkim            #+#    #+#              #
-#    Updated: 2021/12/19 20:55:13 by youkim           ###   ########.fr        #
+#    Updated: 2021/12/20 09:48:40 by youkim           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -38,8 +38,7 @@ PKGS     := engine utils quicksort
 
 engineV    := push_swap engine checks
 utilsV     := opers opers_util visualize
-quicksortV := a_to_b b_to_a smolsort \
-	util_compares util_compares_optimized util_numbers util_quicksort
+quicksortV := quicksort smolsort util_compares util_numbers util_quicksort
 
 # ===== Macros =====
 define choose_modules
