+++ src/engine/push_swap.c
@@ -6,7 +6,7 @@
 /*   By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/11 13:44:15 by youkim            #+#    #+#             */
-/*   Updated: 2021/12/20 09:05:44 by youkim           ###   ########.fr       */
+/*   Updated: 2021/12/20 09:44:17 by youkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,24 +19,13 @@ static void	check_input(const int argc, const char *argv[])
 	// check_duplicate(argc, argv);
 }
 
-void	solve(t_engine *engine)
-{
-	const int	size = engine->a->size;
-
-	if (size <= 3)
-		return ((void)smolsort(engine, STK_A, size));
-	else if (is_sort_complete(engine))
-		return ;
-	partition_initial(engine);
-}
-
 int	main(const int argc, const char *argv[])
 {
 	t_engine	engine;
 
 	check_input(argc, argv);
 	init_engine(&engine, get_input(argc, argv));
-	solve(&engine);
+	quicksort(&engine);
 	yassert(is_sort_complete(&engine), "not sorted!");
 	engine_print_opers(&engine);
 	del_engine(&engine);
